requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
jinja2==3.1.4
pydantic-settings==2.5.2
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
openai==1.53.0
